refactor: Split modules into api/impl for build performance
Split all modules into api (interfaces/contracts) and impl
(implementations) submodules to massively reduce compilation cascades.

## Changes

- **core:fileaccess** → core:fileaccess:api + core:fileaccess:impl
- **core:identity** → core:identity:api + core:identity:impl
- **core:process:search** → core:process:search:api + impl
- **core:process:core-loop** → core:process:core-loop:api + impl
- **agents** → agents:api + agents:impl

## Architecture Benefits

- Implementation changes won't trigger recompilation of dependents
- Clear module boundaries with explicit public APIs
- Faster incremental builds as codebase grows
- Dependencies are more intentional and visible

## Known Issue

Currently experiencing a Gradle KMP circular dependency bug in
metadata compilation tasks for intermediate source sets. This is
a known Gradle issue, not a problem with the module architecture.
