+++ core/identity/api/src/commonMain/kotlin/CoreIdentity.kt
@@ -1,0 +1,8 @@
+package com.brainiac.core.identity
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CoreIdentity(
+    val content: String,
+)
