+++ core/identity/api/src/commonMain/kotlin/CoreIdentityService.kt
@@ -1,0 +1,7 @@
+package com.brainiac.core.identity
+
+interface CoreIdentityService {
+    fun getCoreIdentity(): CoreIdentity
+
+    fun getCoreIdentityContent(): String
+}
