--- core/identity/src/commonMain/kotlin/CoreIdentity.kt
@@ -1,8 +1,0 @@
-package com.brainiac.core.identity
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CoreIdentity(
-    val content: String,
-)
