--- core/identity/src/commonMain/kotlin/CoreIdentityService.kt
@@ -1,9 +1,0 @@
-package com.brainiac.core.identity
-
-import com.brainiac.core.identity.CoreIdentity
-
-interface CoreIdentityService {
-    fun getCoreIdentity(): CoreIdentity
-    
-    fun getCoreIdentityContent(): String
-}
