--- core/identity/src/commonMain/kotlin/DefaultCoreIdentityService.kt
@@ -1,20 +1,0 @@
-package com.brainiac.core.identity
-
-import com.brainiac.core.fileaccess.FileSystemService
-import com.brainiac.core.identity.CoreIdentity
-import okio.Path
-
-class DefaultCoreIdentityService(
-    private val fileSystemService: FileSystemService,
-    private val coreIdentityPath: Path
-) : CoreIdentityService {
-    
-    override fun getCoreIdentity(): CoreIdentity {
-        val content = fileSystemService.read(coreIdentityPath)
-        return CoreIdentity(content)
-    }
-    
-    override fun getCoreIdentityContent(): String {
-        return fileSystemService.read(coreIdentityPath)
-    }
-}
