+++ core/process/core-loop/api/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 kotlin {
     applyDefaultHierarchyTemplate()
+
     jvm()
 
     linuxX64()
@@ -14,8 +15,7 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                // Expose API module to consumers - they only see interfaces
-                api(project(":agents:api"))
+                // Empty - just interface definition
             }
         }
     }
