+++ core/process/core-loop/api/src/commonMain/kotlin/CoreLoopProcess.kt
@@ -1,0 +1,5 @@
+package com.brainiac.core.process
+
+interface CoreLoopProcess {
+    fun processUserPrompt(userPrompt: String): String
+}
