+++ core/process/core-loop/build.gradle.kts
@@ -1,11 +1,9 @@
 plugins {
     alias(libs.plugins.kotlin.multiplatform)
-    alias(libs.plugins.ksp)
-    alias(libs.plugins.mokkery)
-    alias(libs.plugins.kotest)
 }
 
 kotlin {
+    applyDefaultHierarchyTemplate()
     jvm()
 
     linuxX64()
@@ -16,30 +14,9 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                implementation(project(":core:fileaccess"))
-                implementation(project(":core:identity"))
-                implementation(project(":core:process:search"))
-                implementation(libs.kotlinx.datetime)
+                // Expose API module to consumers - they only see interfaces
+                api(project(":core:process:core-loop:api"))
             }
         }
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-                implementation(libs.kotest.framework.engine)
-                implementation(libs.kotest.assertions.core)
-            }
-        }
-        val jvmTest by getting {
-            dependencies {
-                implementation(libs.kotest.runner.junit5)
-            }
-        }
-    }
-}
-
-tasks.named<Test>("jvmTest") {
-    useJUnitPlatform()
-    filter {
-        isFailOnNoMatchingTests = false
     }
 }
