+++ core/process/core-loop/impl/build.gradle.kts
@@ -1,0 +1,47 @@
+plugins {
+    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.ksp)
+    alias(libs.plugins.mokkery)
+    alias(libs.plugins.kotest)
+}
+
+kotlin {
+    applyDefaultHierarchyTemplate()
+    jvm()
+
+    linuxX64()
+    macosX64()
+    macosArm64()
+    mingwX64()
+
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                api(project(":core:process:core-loop:api"))
+                implementation(project(":core:fileaccess:impl"))
+                implementation(project(":core:identity:impl"))
+                implementation(project(":core:process:search:impl"))
+                implementation(libs.kotlinx.datetime)
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+                implementation(libs.kotest.framework.engine)
+                implementation(libs.kotest.assertions.core)
+            }
+        }
+        val jvmTest by getting {
+            dependencies {
+                implementation(libs.kotest.runner.junit5)
+            }
+        }
+    }
+}
+
+tasks.named<Test>("jvmTest") {
+    useJUnitPlatform()
+    filter {
+        isFailOnNoMatchingTests = false
+    }
+}
