+++ core/process/search/api/src/commonMain/kotlin/SearchService.kt
@@ -1,0 +1,7 @@
+package com.brainiac.core.search
+
+import com.brainiac.core.fileaccess.LTMFile
+
+interface SearchService {
+    fun searchLTM(query: String): List<LTMFile>
+}
